# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.AddTelegramRequest:
    properties:
      code:
        type: string
    type: object
  dto.AddTelegramResponse:
    properties:
      telegrams:
        items:
          $ref: '#/definitions/dto.Telegram'
        type: array
    type: object
...
  dto.GetTelegramRequest:
    properties:
      id:
        type: string
    type: object
  dto.GetTelegramResponse:
    properties:
      telegram:
        $ref: '#/definitions/dto.Telegram'
    type: object
  dto.GetTelegramsRequest:
    type: object
  dto.GetTelegramsResponse:
    properties:
      telegrams:
        items:
          $ref: '#/definitions/dto.Telegram'
        type: array
    type: object
  dto.IcePhenomenia:
    properties:
      intensity:
        type: integer
      phenomen:
        type: integer
    type: object
...
  dto.RemoveTelegramsRequest:
    properties:
      id:
        items:
          type: string
        type: array
    type: object
  dto.RemoveTelegramsResponse:
    properties:
      success:
        type: boolean
    type: object
...
  dto.Telegram:
    properties:
      air_temperature:
        type: integer
      average_reservoir_level:
        type: integer
      datetime:
        type: string
      delta_water_level:
        type: integer
      downstream_level:
        type: integer
      group_id:
        type: string
      headwater_level:
        type: integer
      ice_height:
        type: integer
      ice_phenomenia_state:
        type: integer
      ice_phenomenias:
        items:
          $ref: '#/definitions/dto.IcePhenomenia'
        type: array
      id:
        type: string
      inflow:
        type: number
      is_dangerous:
        type: boolean
      post_code:
        type: string
      precipitation_duration:
        type: integer
      precipitation_value:
        type: number
      reservoir_date:
        type: string
      reservoir_volume:
        type: number
      reservoir_water_inflow_date:
        type: string
      reset:
        type: number
      snow_height:
        type: integer
      telegram_code:
        type: string
      water_flow:
        type: number
      water_level_on_time:
        type: integer
      water_level_on20h:
        type: integer
      water_temperature:
        type: number
    type: object
  dto.TransferToSystemRequest:
    type: object
  dto.TransferToSystemResponse:
    properties:
      success:
        type: boolean
    type: object
...
  dto.UpdateTelegramByCodeRequest:
    properties:
      id:
        type: string
      telegram_code:
        type: string
    type: object
  dto.UpdateTelegramByCodeResponse:
    properties:
      telegram:
        $ref: '#/definitions/dto.Telegram'
    type: object
  dto.UpdateTelegramByInfoRequest:
    properties:
      telegram:
        $ref: '#/definitions/dto.Telegram'
    type: object
  dto.UpdateTelegramByInfoResponse:
    properties:
      telegram:
        $ref: '#/definitions/dto.Telegram'
    type: object
...
info:
  contact: {}
  description: This is the HLGateway server API.
  title: HLGateway API
  version: "1"
paths:
...
  /api/add-telegram:
    post:
      consumes:
      - application/json
      description: Add a new telegram
      parameters:
      - description: Add Telegram Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AddTelegramRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AddTelegramResponse'
      summary: Add telegram
      tags:
      - Buffer
...
  /api/get-telegram:
    get:
      consumes:
      - application/json
      description: Get Telegram by id
      parameters:
      - description: Get Telegram Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.GetTelegramRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetTelegramResponse'
      summary: Get Telegram
      tags:
      - Buffer
  /api/get-telegrams:
    get:
      consumes:
      - application/json
      description: Get all Telegrams
      parameters:
      - description: Get Telegrams Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.GetTelegramsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetTelegramsResponse'
      summary: Get Telegrams
      tags:
      - Buffer
...
  /api/remove-telegrams:
    delete:
      consumes:
      - application/json
      description: Remove existing telegrams
      parameters:
      - description: Remove Telegrams Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RemoveTelegramsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RemoveTelegramsResponse'
      summary: Remove telegrams
      tags:
      - Buffer
  /api/transfer-to-system:
    get:
      consumes:
      - application/json
      description: Transfer To System
      parameters:
      - description: Transfer To System Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.TransferToSystemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TransferToSystemResponse'
      summary: Transfer To System
      tags:
      - Buffer
...
  /api/update-telegram-by-code:
    put:
      consumes:
      - application/json
      description: Update info about telegram By Code
      parameters:
      - description: Update Telegram By Code Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTelegramByCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UpdateTelegramByCodeResponse'
      summary: Update Telegram By Code
      tags:
      - Buffer
  /api/update-telegram-by-info:
    put:
      consumes:
      - application/json
      description: Update info about telegram By Info
      parameters:
      - description: Update Telegram By Info Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTelegramByInfoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UpdateTelegramByInfoResponse'
      summary: Update Telegram By Info
      tags:
      - Buffer
...
swagger: "2.0"